+++ opentelemetry-sdk/CHANGELOG.md
@@ -13,13 +13,13 @@ those threads will not be fed back into OTel. Similarly, `SimpleLogProcessor`
 also modified to suppress telemetry before invoking exporters.
 
 - **Feature**: Implemented and enabled cardinality capping for Metrics by
-  default.  
+  default. [#2901](https://github.com/open-telemetry/opentelemetry-rust/pull/2901)
   - The default cardinality limit is 2000 and can be customized using Views.  
   - This feature was previously removed in version 0.28 due to the lack of
     configurability but has now been reintroduced with the ability to configure
     the limit.  
-  - There is ability to configure cardinality limits via Instrument
-    advisory.
+  - There is ability to configure cardinality limits via Instrument 
+    advisory. [#2903](https://github.com/open-telemetry/opentelemetry-rust/pull/2903)
 
 ## 0.29.0