+++ opentelemetry/CHANGELOG.md
@@ -25,7 +25,7 @@ predictable and efficient observability pipelines.
 
 - re-export `tracing` for `internal-logs` feature to remove the need of adding `tracing` as a dependency
 - Added ability to configure cardinality limits via Instrument
-  advisory.
+  advisory. [#2903](https://github.com/open-telemetry/opentelemetry-rust/pull/2903)
 
 ## 0.29.1